+++ src/main/java/com/epam/reportportal/service/LoggingContext.java
@@ -95,7 +95,16 @@ public class LoggingContext {
 	 * Disposes current logging context
 	 */
 	public static void dispose() {
-		ofNullable(getContext()).map(Deque::poll).ifPresent(USED_CONTEXTS::add);
+		int removeFactor = 100;
+		ofNullable(getContext()).map(Deque::poll).ifPresent(context -> {
+			USED_CONTEXTS.add(context);
+			if (context.hashCode() % removeFactor == 0) {
+				USED_CONTEXTS.removeIf(ctx -> {
+					ctx.completables.removeIf(c -> c.test().completions() > 0);
+					return ctx.completables.isEmpty();
+				});
+			}
+		});
 	}
 
 	public static Completable completed() {