+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Cleanup of Completables, by @HardNorth
+- `Utils.getFile(java.net.URI)` method, by @HardNorth
 
 ## [5.3.9]
 ### Fixed